Update settings.py
The ANSI escape sequences for colored text output does not work with the default Windows command prompt CLI unless VT100 emulation is enabled. I added this support for VT100 with Windows CLI directly into the program. Other solutions exist such as python modules or bundling components which modify the default ANSI.sys CLI funtionality on Windows. Ultimately, I used this quick method because I felt that this is the fasted and most strightforward solution for adding the colored text output functionality on Windows.